requirements.txt
flask
flask-mail
flask-wtf
flask-sqlalchemy
flask-migrate
flask-jsglue
tmdbv3api
passlib
email_validator
flask-babel
mysql-connector
tmdbsimple
html5lib
beautifulsoup4
gunicorn
python-slugify
pillow
termcolor
bs4
